requirements.txt
<<<<<<< HEAD
streamlit
google-generativeai
load-dotenv
=======
streamlit
google-ai-generativelanguage 0.4.0
google-api-core              2.15.0
google-auth                  2.25.2
google-generativeai          0.3.2
googleapis-common-protos     1.62.0
translate                    3.6.1
langdetect                   1.0.9
>>>>>>> f44384e (aman aja)
